Move Task.convertToComplete() to TaskUtil
